move loc_insurancemaps app
