+++ ohmg/loc_insurancemaps/management/commands/configure.py
