+++ ohmg/loc_insurancemaps/management/commands/set_cog_symlinks.py
